src/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
python-dotenv==0.19.0 